update unit spawner base mechanic and UI
set up the tile map script to have its variable freely available (most importantly the ground and obstacle layers) to be accessed form other scripts. then implemented a spawn units system to spawn player units on available tiles (not empty, not with a unit already on it, and not an obstacle). finally added a simple mechanic to keep track of units within the spawner and units already on the field and added a temporary UI for it as well.